refactor: split trade insights ai validators
