+++ src/uw_scan/reports/trade_insights_ai/validator_rules/__init__.py
@@ -1,0 +1,1 @@
+"""Rule helpers for Trade Insights AI validation."""
