+++ src/uw_scan/reports/trade_insights_ai/validator_rules/sources.py
@@ -1,0 +1,125 @@
+"""Source-path validation helpers for Trade Insights AI outcomes."""
+
+from __future__ import annotations
+
+import re
+from typing import Any
+
+from uw_scan.models import TradeInsightAiOutcome
+
+_PATH_PART_INDEX_RE = re.compile(r"\[-?\d*\]")
+_SOURCE_PATH_ALIASES = (
+    (
+        "tabs.market_structure.market_structure.stock_history.",
+        "tabs.market_structure.stock_history.",
+    ),
+)
+
+
+def _canonical_source_path(path: str, deterministic_payload: dict[str, Any]) -> str:
+    for invalid_prefix, canonical_prefix in _SOURCE_PATH_ALIASES:
+        if not path.startswith(invalid_prefix):
+            continue
+        candidate = f"{canonical_prefix}{path.removeprefix(invalid_prefix)}"
+        if _path_family_exists(candidate, deterministic_payload):
+            return candidate
+    return path
+
+
+def _path_family_exists(path: str, deterministic_payload: dict[str, Any]) -> bool:
+    parts = [_PATH_PART_INDEX_RE.sub("", p) for p in path.split(".") if p]
+    # Codex sometimes emits dotted-numeric indices (rows.1.spot) instead of
+    # the prompt's bracketed form (rows[N].spot). Both reference the same
+    # field — strip pure-digit segments so the family check accepts either.
+    parts = [p for p in parts if not p.isdigit()]
+    if not parts:
+        return False
+    return _path_parts_exist(deterministic_payload, parts)
+
+
+def _path_parts_exist(node: Any, parts: list[str]) -> bool:
+    if not parts:
+        return True
+    if isinstance(node, list):
+        if not node:
+            return False
+        return any(_path_parts_exist(item, parts) for item in node)
+    part = parts[0]
+    if not isinstance(node, dict) or part not in node:
+        return False
+    return _path_parts_exist(node[part], parts[1:])
+
+
+def _mentions_missing_data(item: Any) -> bool:
+    if not isinstance(item, dict):
+        return False
+    text = " ".join(str(item.get(key) or "") for key in ("label", "value", "note"))
+    return "missing" in text.lower() or "unavailable" in text.lower()
+
+
+def _missing_data_mentions(outcome: TradeInsightAiOutcome, token: str) -> bool:
+    needle = token.lower()
+    return any(needle in item.lower() for item in outcome.missing_data)
+
+
+def _validate_source_path_item(
+    item: Any,
+    deterministic_payload: dict[str, Any],
+    outcome: TradeInsightAiOutcome,
+) -> None:
+    source_path = getattr(item, "source_path", None)
+    item_payload = item.model_dump(mode="json") if hasattr(item, "model_dump") else item
+    if not source_path:
+        if _mentions_missing_data(item_payload):
+            return
+        raise ValueError(
+            "source_path is required unless the item is a missing-data note"
+        )
+    lowered = source_path.lower()
+    # `vanna` and `charm` used to be unavailable in this product; PR #60 made
+    # them first-class fields via exposures_summary + strike_exposures. Allow
+    # references provided the path family resolves below. `short_interest` is
+    # still indirect — only the `short_data` block is forwarded — so it stays
+    # gated on a missing-data acknowledgement.
+    if "short_interest" in lowered and not _missing_data_mentions(
+        outcome, "short_interest"
+    ):
+        raise ValueError(f"unavailable source field referenced: {source_path}")
+    canonical_source_path = _canonical_source_path(source_path, deterministic_payload)
+    if canonical_source_path != source_path:
+        item.source_path = canonical_source_path
+        source_path = canonical_source_path
+    if not _path_family_exists(source_path, deterministic_payload):
+        raise ValueError(f"source_path prefix does not exist: {source_path}")
+
+def _drop_invalid_source_path_in_lenient(
+    item: Any,
+    deterministic_payload: dict[str, Any],
+    outcome: TradeInsightAiOutcome,
+    missing_data: list[str],
+) -> None:
+    """Lenient counterpart to `_validate_source_path_item`: instead of
+    raising on an invalid `source_path`, canonicalize valid ones, and set
+    invalid ones to None while recording the drop in `missing_data`."""
+    source_path = getattr(item, "source_path", None)
+    if not source_path:
+        return
+    lowered = source_path.lower()
+    if "short_interest" in lowered and not _missing_data_mentions(
+        outcome, "short_interest"
+    ):
+        item.source_path = None
+        note = f"source_path dropped (unavailable): {source_path}"
+        if note not in missing_data:
+            missing_data.append(note)
+        return
+    canonical_source_path = _canonical_source_path(source_path, deterministic_payload)
+    if canonical_source_path != source_path:
+        item.source_path = canonical_source_path
+        source_path = canonical_source_path
+    if not _path_family_exists(source_path, deterministic_payload):
+        item.source_path = None
+        note = f"source_path dropped (unknown prefix): {source_path}"
+        if note not in missing_data:
+            missing_data.append(note)
+
